Update use_sbas.md
In the "Estimate Lake Level" section in the first command, we make the SNR files for the year 2020, but the following two gnssir commands attempt to compute reflector heights for the year 2021, so they end with an error because there are no SNR files for that year. I changed them to 2020 and my plots look identical to the ones there, so it is just a typo and those two commands should say 2020 instead of 2021.