Update gnssir_v2.py
Add a check for length of x -- this fails for x of length 0, and kicks out a loop of e.g., "gnssir trda 2023 1 -doy_end 365 -newarcs T -plt True". Chasing this down leads to finding sometimes an empty 'x' array.

This seems to fix the issue - will run through the rest of the day/following days in my tests.