Add local file import
I believe when I used gnssrefl.rinex2snr import, it used the default
library which I had downloaded using pip which did not expect the
par argument. When I load the local file manually, it seems to work
fine. I don't know what a more appropriate fix for this is.